forecasting_dashboard_app/requirements.txt
streamlit==1.29.0
pandas==2.1.3
plotly==5.18.0
duckdb==0.9.2
python-dotenv==1.0.0
typing-extensions==4.8.0